requirements.txt
chardet
tensorflow
scikit-learn
matplotlib
numpy
pandas
keras
scipy
pillow
torchvision
flask
